requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29
six==1.14.0

# data science
numpy==1.18.5
pandas==1.0.4
scikit-learn==0.22

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow==1.16.0
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

cmdstanpy

fastapi
uvicorn

